Remove const declarations and actually save the new target
The segfault previously described appears to have been due to the use of `const` in the ranged for-loop, and the mismatched type in the reinterpret_cast.
A second issue was that the wrong pointer was being used to save the result of the target acquisition routine.
Third, existing code was reviewed to ensure that the update would be lasting (e.g. targetShip must be updated, not only cachedTarget)