Work in progress toward introducing multiple destinations
This is itself the first step toward a patrol functionality.

Another key feature is that the destinationQueue counter fixes another
issue. Without it, try this: 1) destination + !staying, 2) begin
mission, let the NPC reach you, 3) take it to some new location. 4)
Quit, relaunch. You'll see it decides to go back to its 'destination',
not remembering it already did that.

(destinationQueue will need to be saved at ship level. That's one of my
next jobs.)