Refueling for MoveIndependent
NPCs that use MoveIndependent will benefit from both wormhole routing logic, and refuel if necessary or will be needed but unavailable after the next jump.
Since NPCs with a waypoint or patrol behave a little differently than normal NPCs (they don't get normal "should I land?" consideration), the check for !ship.JumpsRemaining() is needed. If fuel is unavailable in a system, the NPCs will ask other ships - but only if the player is present will there actually be other ships that can be asked for help.