Added automatic waypoint/stopover generation from filters
In addition to the existing one/multi line definition methods constructed previously, waypoints and stopovers are now randomly generated from the normal "filter" syntax.

Patrols can also be automatically generated between the mission's origin and destination with the sole keyword "patrol"

Fixed a bug in FollowOrders, for which if a system was inaccessible, an unhandled nullptr was generated.
 - Inaccessible waypoints will now generate a non-action, and the ship will not have hyperspace target updated.

Added invalid waypoint handling at the ship level, with EraseWaypoint,
 - Removes any inaccessible waypoints from the ship's travel directives.
 - The whole waypoint list is stored at the NPC level, as they may be valid for some ships in the NPC block.
 - This can result in an NPC resuming a patrol incorrectly, as its index will be relative to a different size container.