Mission success/failure handling for landed ships
Destroy and land interact correctly now, i.e., if you need a ship to
land and it's destroyed, the mission will fail. A landed ship will also
fail a mission if you still need to board, scan, etc.

Was able to simplify away the mustLiveFor bitmask, as the target must be
alive for any succeedIf to be done.