Vomitting my notes out onto the world
deep missions.txt contains a bunch of notes and skeleton missions for
how the Deep missions will be written and play out for the player to be
able to unlock the Bactrian.

Right now this is just a load of vomit that might be hard to follow.
I'll start filling in the holes tomorrow, starting (although out of
order) with the Deep: Remnant string of missions (since I already have
in mind how the entire string will go), then working my way back to
Deep: TMBR and Deep: Convoy.

A rewrite of Deep: Star Queen Rescue will be the final thing.