Changing the to offer conditions on some of the "start" Deep missions
The same number of jobs need to be done to advance to the "middle" Deep
mission, but now you can get some of the "start" Deep missions sooner.

The jobs you need to do relate to the missions. e.g. Deep: TMBR requires
the scientist related jobs because you're dealing with scientists. The
Deep: Convoy missions require mystery cube jobs because those cubes will
relate to the missions in some way. (This is also why Deep: Mystery
Cubes requires one convoy mission to be done.)