Removing redundant condition
Part of starting the Deep: Remnant string involves having either 3
convoys done or finishing the Deep: TMBR string. Therefore, placing it
in an or structure with those exact two conditions is redudnant since
the precense of the has "Deep: Remnant Research: done" condition means
that at least of the other two conditions must already be true as well.