Housekeeping + an uncomplete Deep: Singularity skeleton
I really have no idea what I'm going to do with the singularity missions
to make them different from the rest of the missions. None of the
missions so far involve transporting cargo so it might involve that, but
even then I don't know how to tie that in to the singularity.

Also, my deadline for finishing this up is this Wednesday, so I'm going
to try working on this all day tomorrow and hopefully get Deep:
Singularity completely done and start on rewriting Deep: Scientist
Rescue.