Consider the next refuel stop when following the player's travel plan.
When an escort is jumping with the player, it will decide to refuel based
on the next system with fuel when following the player's travel plan.

It supports travel plans that are impossible to follow directly, like when
the player has a jump drive and the escort has a hyperdrive.